requirements.txt
absl-py==2.0.0
aiosignal==1.3.1
ale-py==0.8.1
astunparse==1.6.3
attrs==23.1.0
backports.functools-lru-cache==1.6.3
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==2.0.10
click==8.1.7
cloudpickle==2.0.0
comm==0.1.0
contourpy==0.5.3
cycler==0.10.0
Cython==0.29.36
debugpy==1.5.0
decorator==4.4.2
dm-tree==0.1.8
etils==1.5.2
exceptiongroup==0.31.0
executing==0.9.0
fasteners==0.19.1
filelock==3.0.12
flatbuffers==23.5.26
fonttools==5.0.0
frozenlist==1.4.0
fsspec==2023.10.0
gast==0.5.4
glfw==2.6.2
gmpy2==2.1.0b6
google-auth==2.23.4
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.59.2
gymnasium==0.28.1
h5py==3.10.0
idna==3.3
imageio==2.32.0
imageio-ffmpeg==0.4.9
importlib-metadata==4.10.2
importlib-resources==6.1.1
ipykernel==6.7.0
ipython==8.3.0
jax==0.4.20
jax-jumpy==0.2.18
jaxlib==0.1.79
jedi==0.18.0
Jinja2==3.1.3
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
jupyter_client==7.1.2
jupyter_core==4.9.2
keras==2.15.0
kiwisolver==1.3.2
lazy_loader==0.3
libclang==16.0.6
lz4==4.3.2
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdurl==0.1.2
ml-dtypes==0.2.0
moviepy==1.0.3
mpmath==1.2.1
msgpack==1.0.7
nest-asyncio==1.5.4
networkx==2.7.3
numpy==1.26.0
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==22.0
pandas==1.3.3
parso==0.8.3
pexpect==4.10.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.0.2
proglog==0.1.10
prompt-toolkit==3.0.28
protobuf==4.23.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.3.6
pyarrow==6.0.1
pyasn1==0.5.0
pyasn1-modules==0.2.7
pycparser==2.21
pygame==2.5.2
Pygments==2.10.0
PyOpenGL==3.1.7
pyparsing==3.0.0
PySocks==1.7.1
python-dateutil==2.8.3
pytz==2023.3.post1
PyYAML==6.0
pyzmq==22.3.0
ray==2.8.0
referencing==0.31.0
requests==2.26.0
requests-oauthlib==1.3.1
rich==13.7.0
-e git+ssh://git@github.com/Alexyskoutnev/RL_Pizza_Delivery.git@cdbc0ddf887fea19f4ae6d9b4d0beeffbe74b135#egg=RL_Pizza_Delivery
rpds-py==0.12.0
rsa==4.9
scikit-image==0.22.0
scipy==1.11.3
Shimmy==0.2.1
six==1.16.0
stack-data==1.13.0
swig==4.1.1.post0
sympy==1.11
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
termcolor==1.1.0
tifffile==2023.9.26
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.1
tqdm==4.66.1
traitlets==5.1.1
typer==0.9.0
typing_extensions==4.0.1
tzdata==2023.3
urllib3==1.26.9
wcwidth==0.2.5